fix(reference): narrow live-proof bootstrap path
