+++ meshlink-reference/src/commonMain/kotlin/ch/trancee/meshlink/reference/navigation/SessionAwarePlatformServices.kt
@@ -1,13 +1,13 @@
 package ch.trancee.meshlink.reference.navigation
 
 import ch.trancee.meshlink.reference.meshlink.ReferenceMeshLinkController
-import ch.trancee.meshlink.reference.platform.PlatformServices
+import ch.trancee.meshlink.reference.platform.LiveProofPlatformServices
 import ch.trancee.meshlink.reference.session.ReferenceSessionController
 
 internal class SessionAwarePlatformServices(
-    private val delegate: PlatformServices,
+    private val delegate: LiveProofPlatformServices,
     private val sessionController: ReferenceSessionController,
-) : PlatformServices by delegate {
+) : LiveProofPlatformServices by delegate {
     override val meshLinkController: ReferenceMeshLinkController
         get() = sessionController
